Moved the code repository in a separate module
