+++ configs/common/node.json
@@ -2,7 +2,7 @@
     "forum": {
         "name": "forum",
         "logger": {
-          "level": "debug"
+          "level": "info"
         },
         "port": 3000,
         "template": {
